--- frontend/src/constants/workTimePx.ts
@@ -1,3 +1,0 @@
-const WORK_TIME_DEFAULT_X_COORDINATE = 16;
-
-export { WORK_TIME_DEFAULT_X_COORDINATE };
