+++ frontend/src/lib/dnd/utils/isFullyOverlappedUtil.ts
