+++ frontend/src/lib/dnd/utils/snapToGridUtil.ts
